// src/app/models/order.model.ts
export interface OrderItem {
    productId: number;
    name: string;
    quantity: number;
    price: number;
}

export interface Order {
    id?: number;
    userId: number;
    total: number;
    status: 'pending' | 'processing' | 'delivered' | 'cancelled';
    date: string;
    items: OrderItem[];
}